requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
bcrypt==4.1.1
requests==2.31.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
